third_party/blink/web_tests/fast/webmcp/disabled-web-security.https.html

<!DOCTYPE html>
<title>WebMCP: WebMCP is allowed when web security is disabled, even without origin keying</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
if (window.testRunner) {
  testRunner.overridePreference("WebKitWebSecurityEnabled", false);
}

test(() => {
  assert_false(self.originAgentCluster, "Document should not be origin-keyed");

  const tool = {
    name: 'test_tool',
    description: 'a test tool',
    execute: () => 'success',
  };

  // This should NOT throw SecurityError because web security is disabled.
  document.modelContext.registerTool(tool);
}, 'WebMCP is allowed when web security is disabled, even if not origin-keyed');
</script>
